# Faster conntrack parser