--- a/probe/endpoint/conntrack.go
+++ b/probe/endpoint/conntrack.go
@@ -26,12 +26,6 @@ const (
 	destroyType = "DESTROY"
 )
 
-var (
-	destroyTypeB = []byte(destroyType)
-	assured      = []byte("[ASSURED] ")
-	unreplied    = []byte("[UNREPLIED] ")
-)
-
 type layer3 struct {
 	SrcIP string
 	DstIP string