--- a/probe/endpoint/conntrack.go
+++ b/probe/endpoint/conntrack.go
@@ -215,6 +215,118 @@ func (c *conntrackWalker) run() {
 	}
 }
 
+// decodeFlowKeyValues parses the key-values from a conntrack line and updates the flow
+// It only considers the following key-values:
+// src=127.0.0.1 dst=127.0.0.1 sport=58958 dport=6784 src=127.0.0.1 dst=127.0.0.1 sport=6784 dport=58958 id=1595499776
+// Keys can be present twice, so the order is important.
+// Conntrack could add other key-values such as secctx=, packets=, bytes=. Those are ignored.
+func decodeFlowKeyValues(s *Scanner, f *flow) error {
+	for {
+		var err error
+		key := keyNone
+		tok := s.scan()
+		if tok == NEWLINE || tok == EOF {
+			break
+		} else if tok == IDENT {
+			key = s.lastKeyType()
+		}
+		if key == keyNone {
+			s.skipToWhitespace()
+			continue
+		}
+		if err := s.mustBe(EQUALS); err != nil {
+			return err
+		}
+		if tok = s.scan(); tok != NUMERIC && tok != IDENT {
+			return s.errorExpected(IDENT, tok)
+		}
+
+		firstTupleSet := f.Original.Layer4.DstPort != 0
+		switch key {
+		case keySrc:
+			if !firstTupleSet {
+				f.Original.Layer3.SrcIP = s.lastValue()
+			} else {
+				f.Reply.Layer3.SrcIP = s.lastValue()
+			}
+
+		case keyDst:
+			if !firstTupleSet {
+				f.Original.Layer3.DstIP = s.lastValue()
+			} else {
+				f.Reply.Layer3.DstIP = s.lastValue()
+			}
+
+		case keySport:
+			if !firstTupleSet {
+				f.Original.Layer4.SrcPort, err = s.lastInt()
+			} else {
+				f.Reply.Layer4.SrcPort, err = s.lastInt()
+			}
+
+		case keyDport:
+			if !firstTupleSet {
+				f.Original.Layer4.DstPort, err = s.lastInt()
+			} else {
+				f.Reply.Layer4.DstPort, err = s.lastInt()
+			}
+
+		case keyID:
+			f.Independent.ID, err = s.lastInt64()
+		}
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func decodeStreamedFlow(s *Scanner) (flow, error) {
+	var f flow
+	// Examples:
+	// " [UPDATE] udp      17 29 src=192.168.2.100 dst=192.168.2.1 sport=57767 dport=53 src=192.168.2.1 dst=192.168.2.100 sport=53 dport=57767"
+	// "    [NEW] tcp      6 120 SYN_SENT src=127.0.0.1 dst=127.0.0.1 sport=58958 dport=6784 [UNREPLIED] src=127.0.0.1 dst=127.0.0.1 sport=6784 dport=58958 id=1595499776"
+	// " [UPDATE] tcp      6 120 TIME_WAIT src=10.0.2.15 dst=10.0.2.15 sport=51154 dport=4040 src=10.0.2.15 dst=10.0.2.15 sport=4040 dport=51154 [ASSURED] id=3663628160"
+	// " [DESTROY] tcp      6 src=172.17.0.1 dst=172.17.0.1 sport=34078 dport=53 src=172.17.0.1 dst=10.0.2.15 sport=53 dport=34078 id=3668417984" (note how the timeout and state field is missing)
+
+	// First token should be a square bracket for the protocol.
+	tok := s.scan()
+	if tok == EOF {
+		return f, io.EOF
+	} else if tok != LSQUARE {
+		return f, s.errorExpected(LSQUARE, tok)
+	}
+	if err := s.mustBe(IDENT); err != nil {
+		return f, err
+	}
+	f.Type = s.lastSymbol()
+	if err := s.mustBe(RSQUARE); err != nil {
+		return f, err
+	}
+
+	if err := s.mustBe(IDENT); err != nil {
+		return f, err
+	}
+	f.Original.Layer4.Proto = s.lastSymbol()
+
+	s.scan() // unused
+	if f.Type != destroyType {
+		s.scan() // unused
+		if err := s.mustBe(IDENT); err != nil {
+			return f, err
+		}
+		f.Independent.State = s.lastSymbol()
+	}
+
+	err := decodeFlowKeyValues(s, &f)
+	if err != nil {
+		return flow{}, fmt.Errorf("Error parsing streamed flow: %v ", err)
+	}
+
+	f.Reply.Layer4.Proto = f.Original.Layer4.Proto
+	return f, nil
+}
+
 func existingConnections(conntrackWalkerArgs []string) ([]flow, error) {
 	args := append([]string{"-L", "-o", "id", "-p", "tcp"}, conntrackWalkerArgs...)
 	cmd := exec.Command("conntrack", args...)
@@ -245,6 +357,45 @@ func existingConnections(conntrackWalkerArgs []string) ([]flow, error) {
 		result = append(result, f)
 	}
 	return result, nil
+}
+
+func decodeDumpedFlow(s *Scanner) (flow, error) {
+	var f flow
+	var tok Token
+
+	// Examples with different formats:
+	// With SELinux, there is a "secctx="
+	// After "sudo sysctl net.netfilter.nf_conntrack_acct=1", there is "packets=" and "bytes="
+	//
+	// "tcp      6 431997 ESTABLISHED src=10.32.0.1 dst=10.32.0.1 sport=50274 dport=4040 src=10.32.0.1 dst=10.32.0.1 sport=4040 dport=50274 [ASSURED] mark=0 use=1 id=407401088"
+	// "tcp      6 431998 ESTABLISHED src=10.0.2.2 dst=10.0.2.15 sport=49911 dport=22 src=10.0.2.15 dst=10.0.2.2 sport=22 dport=49911 [ASSURED] mark=0 use=1 id=2993966208"
+	// "tcp      6 108 ESTABLISHED src=172.17.0.5 dst=172.17.0.2 sport=47010 dport=80 src=172.17.0.2 dst=172.17.0.5 sport=80 dport=47010 [ASSURED] mark=0 secctx=system_u:object_r:unlabeled_t:s0 use=1 id=4001098880"
+	// "tcp      6 431970 ESTABLISHED src=192.168.35.116 dst=216.58.213.227 sport=49862 dport=443 packets=11 bytes=1337 src=216.58.213.227 dst=192.168.35.116 sport=443 dport=49862 packets=8 bytes=716 [ASSURED] mark=0 secctx=system_u:object_r:unlabeled_t:s0 use=1 id=943643840"
+	// First token should be the protocol.
+	tok = s.scan()
+	if tok == EOF {
+		return f, io.EOF
+	} else if tok != IDENT {
+		return f, s.errorExpected(IDENT, tok)
+	}
+	f.Original.Layer4.Proto = s.lastSymbol()
+
+	// two unused fields
+	s.scan()
+	s.scan()
+
+	if err := s.mustBe(IDENT); err != nil {
+		return f, err
+	}
+	f.Independent.State = s.lastSymbol()
+
+	err := decodeFlowKeyValues(s, &f)
+	if err != nil {
+		return flow{}, fmt.Errorf("Error parsing dumped flow: %v ", err)
+	}
+
+	f.Reply.Layer4.Proto = f.Original.Layer4.Proto
+	return f, nil
 }
 
 func (c *conntrackWalker) stop() {