Profiling indicated that conntrack parsing was around 20% of the CPU time of the probe in Weave Cloud dev, so I wrote a scanner/parser instead of relying on `sscanf` and `split`.

Test run decoding a `conntrack -L` dump with 27,000 lines:

Before:
```
ok  	github.com/weaveworks/scope/probe/endpoint	19.174s
ok  	github.com/weaveworks/scope/probe/endpoint	17.206s
ok  	github.com/weaveworks/scope/probe/endpoint	17.377s
ok  	github.com/weaveworks/scope/probe/endpoint	18.486s
```

After
```
ok  	github.com/weaveworks/scope/probe/endpoint	3.817s
ok  	github.com/weaveworks/scope/probe/endpoint	4.411s
ok  	github.com/weaveworks/scope/probe/endpoint	3.702s
ok  	github.com/weaveworks/scope/probe/endpoint	4.022s
```

3.7/17.2=22%

Individual commits left for now in case I need to remember how I got here, but I would squash before merging. I don't recommend reading the individual commits for review.