@2opremio 2017-05-02
Thanks @bboreham ! I shall take a look tomorrow 

It seems you have put quite a bit of work into this, wouldn't it have been better spent ditching the text parser and finishing https://github.com/vishvananda/netlink/issues/171 ? After https://github.com/vishvananda/netlink/pull/213 I think it wouldn't had taken too long.