@2opremio 2017-05-03
Done: https://github.com/weaveworks/scope/issues/2495

@bboreham Mind fixing the linter problems?