@bboreham 2020-01-13
We moved to using netlink in #3298.